requirements.txt
numpy>=1.21.2
pandas>=1.3.3
scipy>=1.7.1
scikit-learn>=0.24.2
matplotlib>=3.4.3
seaborn>=0.11.2
snakemake>=7.0.0
biopython>=1.79
rpy2>=3.4.5
pyyaml>=5.4.1